# http/openapi.gen.go
//go:build go1.22

// Package http provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package http

import (
	"context"
	"fmt"
	"net/http"
	"time"

	"github.com/oapi-codegen/runtime"
)

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// AddFavouriteRequest defines model for AddFavouriteRequest.
type AddFavouriteRequest struct {
	Audience      *AudienceData `json:"audience,omitempty"`
	Chart         *ChartData    `json:"chart,omitempty"`
	Description   *string       `json:"description,omitempty"`
	Insight       *InsightData  `json:"insight,omitempty"`
	SourceAssetId *string       `json:"source_asset_id,omitempty"`
	Type          string        `json:"type"`
}

// AssetDeletedRequest defines model for AssetDeletedRequest.
type AssetDeletedRequest struct {
	AssetId string `json:"asset_id"`
}

// AssetDeletedResponse defines model for AssetDeletedResponse.
type AssetDeletedResponse struct {
	Removed *int `json:"removed,omitempty"`
}

// AudienceData defines model for AudienceData.
type AudienceData struct {
	AgeGroups             *string  `json:"age_groups,omitempty"`
	BirthCountry          *string  `json:"birth_country,omitempty"`
	Gender                *string  `json:"gender,omitempty"`
	HoursSocialMediaDaily *float32 `json:"hours_social_media_daily,omitempty"`
	PurchasesLastMonth    *int     `json:"purchases_last_month,omitempty"`
}

// ChartData defines model for ChartData.
type ChartData struct {
	Data       interface{} `json:"data,omitempty"`
	Title      *string     `json:"title,omitempty"`
	XAxisTitle *string     `json:"x_axis_title,omitempty"`
	YAxisTitle *string     `json:"y_axis_title,omitempty"`
}

// Error defines model for Error.
type Error struct {
	Error *struct {
		Code    *string `json:"code,omitempty"`
		Message *string `json:"message,omitempty"`
	} `json:"error,omitempty"`
}

// FavouriteListItem defines model for FavouriteListItem.
type FavouriteListItem struct {
	CreatedAt     *time.Time `json:"created_at,omitempty"`
	Description   *string    `json:"description,omitempty"`
	Id            *string    `json:"id,omitempty"`
	SourceAssetId *string    `json:"source_asset_id,omitempty"`
	Type          *string    `json:"type,omitempty"`
	UpdatedAt     *time.Time `json:"updated_at,omitempty"`
}

// FavouriteResponse defines model for FavouriteResponse.
type FavouriteResponse struct {
	Audience      *AudienceData `json:"audience,omitempty"`
	Chart         *ChartData    `json:"chart,omitempty"`
	CreatedAt     *time.Time    `json:"created_at,omitempty"`
	Description   *string       `json:"description,omitempty"`
	Id            *string       `json:"id,omitempty"`
	Insight       *InsightData  `json:"insight,omitempty"`
	SourceAssetId *string       `json:"source_asset_id,omitempty"`
	Type          *string       `json:"type,omitempty"`
	UpdatedAt     *time.Time    `json:"updated_at,omitempty"`
}

// InsightData defines model for InsightData.
type InsightData struct {
	Text *string `json:"text,omitempty"`
}

// IssueTokenRequest defines model for IssueTokenRequest.
type IssueTokenRequest struct {
	UserId string `json:"user_id"`
}

// IssueTokenResponse defines model for IssueTokenResponse.
type IssueTokenResponse struct {
	// ExpiresIn Seconds until expiry
	ExpiresIn *int    `json:"expires_in,omitempty"`
	Token     *string `json:"token,omitempty"`
}

// ListFavouritesResponse defines model for ListFavouritesResponse.
type ListFavouritesResponse struct {
	Items  *[]FavouriteListItem `json:"items,omitempty"`
	Limit  *int                 `json:"limit,omitempty"`
	Offset *int                 `json:"offset,omitempty"`
	Total  *int                 `json:"total,omitempty"`
}

// UpdateDescriptionRequest defines model for UpdateDescriptionRequest.
type UpdateDescriptionRequest struct {
	Description *string `json:"description,omitempty"`
}

// VersionResponse API version (no auth required).
type VersionResponse struct {
	Api     *string `json:"api,omitempty"`
	Version *string `json:"version,omitempty"`
}

// AssetID defines model for AssetID.
type AssetID = string

// UserID defines model for UserID.
type UserID = string

// BadRequest defines model for BadRequest.
type BadRequest = Error

// Forbidden defines model for Forbidden.
type Forbidden = Error

// MethodNotAllowed defines model for MethodNotAllowed.
type MethodNotAllowed = Error

// NotFound defines model for NotFound.
type NotFound = Error

// Unauthorized defines model for Unauthorized.
type Unauthorized = Error

// ListFavouritesParams defines parameters for ListFavourites.
type ListFavouritesParams struct {
	Limit  *int `form:"limit,omitempty" json:"limit,omitempty"`
	Offset *int `form:"offset,omitempty" json:"offset,omitempty"`
}

// IssueTokenJSONRequestBody defines body for IssueToken for application/json ContentType.
type IssueTokenJSONRequestBody = IssueTokenRequest

// AssetDeletedJSONRequestBody defines body for AssetDeleted for application/json ContentType.
type AssetDeletedJSONRequestBody = AssetDeletedRequest

// AddFavouriteJSONRequestBody defines body for AddFavourite for application/json ContentType.
type AddFavouriteJSONRequestBody = AddFavouriteRequest

// UpdateFavouriteDescriptionJSONRequestBody defines body for UpdateFavouriteDescription for application/json ContentType.
type UpdateFavouriteDescriptionJSONRequestBody = UpdateDescriptionRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Issue JWT
	// (POST /auth/token)
	IssueToken(w http.ResponseWriter, r *http.Request)
	// Internal event – asset deleted (optional X-Internal-Secret)
	// (POST /internal/events/asset-deleted)
	AssetDeleted(w http.ResponseWriter, r *http.Request)
	// List favourites (paginated)
	// (GET /users/{userID}/favourites)
	ListFavourites(w http.ResponseWriter, r *http.Request, userID UserID, params ListFavouritesParams)
	// Add favourite
	// (POST /users/{userID}/favourites)
	AddFavourite(w http.ResponseWriter, r *http.Request, userID UserID)
	// Remove favourite
	// (DELETE /users/{userID}/favourites/{assetID})
	RemoveFavourite(w http.ResponseWriter, r *http.Request, userID UserID, assetID AssetID)
	// Get one favourite
	// (GET /users/{userID}/favourites/{assetID})
	GetFavourite(w http.ResponseWriter, r *http.Request, userID UserID, assetID AssetID)
	// Update favourite description
	// (PATCH /users/{userID}/favourites/{assetID})
	UpdateFavouriteDescription(w http.ResponseWriter, r *http.Request, userID UserID, assetID AssetID)
	// API version
	// (GET /version)
	GetVersion(w http.ResponseWriter, r *http.Request)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// IssueToken operation middleware
func (siw *ServerInterfaceWrapper) IssueToken(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.IssueToken(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// AssetDeleted operation middleware
func (siw *ServerInterfaceWrapper) AssetDeleted(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.AssetDeleted(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListFavourites operation middleware
func (siw *ServerInterfaceWrapper) ListFavourites(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "userID" -------------
	var userID UserID

	err = runtime.BindStyledParameterWithOptions("simple", "userID", r.PathValue("userID"), &userID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "userID", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	// Parameter object where we will unmarshal all parameters from the context
	var params ListFavouritesParams

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameter("form", true, false, "limit", r.URL.Query(), &params.Limit)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "limit", Err: err})
		return
	}

	// ------------- Optional query parameter "offset" -------------

	err = runtime.BindQueryParameter("form", true, false, "offset", r.URL.Query(), &params.Offset)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "offset", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListFavourites(w, r, userID, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// AddFavourite operation middleware
func (siw *ServerInterfaceWrapper) AddFavourite(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "userID" -------------
	var userID UserID

	err = runtime.BindStyledParameterWithOptions("simple", "userID", r.PathValue("userID"), &userID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "userID", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.AddFavourite(w, r, userID)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// RemoveFavourite operation middleware
func (siw *ServerInterfaceWrapper) RemoveFavourite(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "userID" -------------
	var userID UserID

	err = runtime.BindStyledParameterWithOptions("simple", "userID", r.PathValue("userID"), &userID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "userID", Err: err})
		return
	}

	// ------------- Path parameter "assetID" -------------
	var assetID AssetID

	err = runtime.BindStyledParameterWithOptions("simple", "assetID", r.PathValue("assetID"), &assetID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "assetID", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.RemoveFavourite(w, r, userID, assetID)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetFavourite operation middleware
func (siw *ServerInterfaceWrapper) GetFavourite(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "userID" -------------
	var userID UserID

	err = runtime.BindStyledParameterWithOptions("simple", "userID", r.PathValue("userID"), &userID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "userID", Err: err})
		return
	}

	// ------------- Path parameter "assetID" -------------
	var assetID AssetID

	err = runtime.BindStyledParameterWithOptions("simple", "assetID", r.PathValue("assetID"), &assetID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "assetID", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetFavourite(w, r, userID, assetID)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// UpdateFavouriteDescription operation middleware
func (siw *ServerInterfaceWrapper) UpdateFavouriteDescription(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "userID" -------------
	var userID UserID

	err = runtime.BindStyledParameterWithOptions("simple", "userID", r.PathValue("userID"), &userID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "userID", Err: err})
		return
	}

	// ------------- Path parameter "assetID" -------------
	var assetID AssetID

	err = runtime.BindStyledParameterWithOptions("simple", "assetID", r.PathValue("assetID"), &assetID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "assetID", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.UpdateFavouriteDescription(w, r, userID, assetID)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetVersion operation middleware
func (siw *ServerInterfaceWrapper) GetVersion(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetVersion(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of http.ServeMux.
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	ServeHTTP(w http.ResponseWriter, r *http.Request)
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc("POST "+options.BaseURL+"/auth/token", wrapper.IssueToken)
	m.HandleFunc("POST "+options.BaseURL+"/internal/events/asset-deleted", wrapper.AssetDeleted)
	m.HandleFunc("GET "+options.BaseURL+"/users/{userID}/favourites", wrapper.ListFavourites)
	m.HandleFunc("POST "+options.BaseURL+"/users/{userID}/favourites", wrapper.AddFavourite)
	m.HandleFunc("DELETE "+options.BaseURL+"/users/{userID}/favourites/{assetID}", wrapper.RemoveFavourite)
	m.HandleFunc("GET "+options.BaseURL+"/users/{userID}/favourites/{assetID}", wrapper.GetFavourite)
	m.HandleFunc("PATCH "+options.BaseURL+"/users/{userID}/favourites/{assetID}", wrapper.UpdateFavouriteDescription)
	m.HandleFunc("GET "+options.BaseURL+"/version", wrapper.GetVersion)

	return m
}